WIP - updating user
